HACK: shim out definitions of GLdouble and GLclampd for iOS
It appears that these types are not defined in the iOS opengl types headers,
however we use them in our definitions of OpenGL functions in
`GLContextSymbols.h` for fDepthRange, fClearDepth, and fLoadMatrixd.

Without adding a defintiion for these symbols, the GLContextSymbols.h file
won't build, so this patch adds code to ensure that they are defined on iOS.
This is a hack and should be fixed in a better way before merging to central.