Get MacIOSurface building on iOS
This required changings some files from .cpp to .mm, which appears to have
caused a large number of clang-format formatting changes. This is required
because GLContextEAGL.h is an objective-c++ header, unlike GLContextCGL.h.

There is also an unfortunate amount of duplication between
EAGLTexImageIOSurface2D and CGLTexImageIOSurface2D, for quite complex code,
however cleaning this up was left out of scope for now. We'll probably want to
unify the codepaths together more in the future.

Finally, the `texImageIOSurface:` call used for `EAGLTexImageIOSurface2D`
appears to not be available in iOS simulator headers, only real iOS headers. It
seems like ANGLE has a workaround of some kind which we may want to adapt in
the future, however it wasn't required to get basic rendering working.