Forcibly disable e10s on iOS devices
Due to the strict iOS sandbox, it is currently not possible to start child
processes on iOS. To allow some basic testing on the device, forcibly disable
e10s on these devices for now.

e10s will still be enabled when run in the simulator, where it can be supported.