Hackily support scaling and rotation on iOS
This is a temporary solution. Once we've figured out how we actually want to
embed Gecko using a UIKit widget, we'll stop using a global view controller and
the deprecated applicationFrame member for sizing, as we'll likely be instead
basing the sizing off of UIKit View layout-based sizing constraints instead, as
driven by the embedder application.

This logic also more explicitly encodes the current nsWindow assumption that
only a single visible toplevel window will be used.