Bug 1859357 - Part 3: Replace ExampleBrowser with a project using the new API, r=glandium!
This is a project with very similar functionality to the existing
ExampleBrowser project, however built using the new GeckoView-style embedding
API introduced in the previous part. Like the ExampleBrowser project, this is
built using xcode rather than the normal build system, however basic building,
signing & packaging is now performed automatically by the Makefile when
building on a macOS host.

In another patch, I have a set-up which builds the application bundle manually
within the Makefile, avoiding any xcode dependency. Unfortunately this makes it
more difficult to run on a real iOS device due to the need for manual signing
and provisioning profiles.

This depends on the new `./mach environment --format topobjdir` command which
was introduced in part 1 to allow the xcode project to also be built & run
directly, rather than exclusively through our build system.

Differential Revision: https://phabricator.services.mozilla.com/D191102